Debugging menu

+++ test3.py
@@ -120,6 +120,13 @@ def rotate_press(state):
             Menu_submenu_pointer = 0
             Menu_parent = Menu_menu_pointer
         else:
+            if ((Menu_menu_pointer == 3) and (Menu_submenu_pointer == 1)):
+                os.system("shutdown")
+                Menu_menu_pointer = 0
+                Menu_submenu_pointer = 0
+                Menu_parent = 0
+                Menu_uroven = 0
+                return
             if ((Menu_menu_pointer == 3) and (Menu_submenu_pointer == 0)):
                 os.system("reboot")
                 Menu_menu_pointer = 0
